Split KSC logic to controller and reconciler
