+++ src/operator/controllers/kafka_server_config_reconcilers/kafka_server_config_controller_test.go
@@ -1,4 +1,4 @@
-package controllers
+package kafka_server_config_reconcilers
 
 import (
 	"context"
